# docs/swagger.yaml
basePath: /api
definitions:
  models.Pomodoro:
    properties:
      created_at:
        type: string
      date:
        type: string
      duration:
        type: integer
      id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Task:
    properties:
      created_at:
        type: string
      deadline:
        type: string
      description:
        type: string
      id:
        type: integer
      is_complete:
        type: boolean
      priority:
        description: 0=rendah, 1=sedang, 2=tinggi,3=paling tinggi
        type: integer
      title:
        type: string
      todo:
        $ref: '#/definitions/models.Todo'
      todo_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Todo:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_complete:
        type: boolean
      priority:
        description: 0=rendah, 1=sedang, 2=tinggi,3=paling tinggi
        type: integer
      progress:
        type: integer
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      google_id:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      picture:
        type: string
      pomodoros:
        items:
          $ref: '#/definitions/models.Pomodoro'
        type: array
      provider:
        type: string
      todos:
        items:
          $ref: '#/definitions/models.Todo'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
host: api.tascaid.com
info:
  contact: {}
  description: API untuk aplikasi Tasca A4 AGILE menggunakan gin
  title: Tag Tasca API
  version: "1.0"
paths:
  /api/calendar/day/{date}:
    get:
      description: Get all tasks for a specific date
      parameters:
      - description: Date (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: 'error: Format tanggal tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Gagal mengambil data'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get tasks by date
      tags:
      - calendar
...
  /api/tasks/complete:
    get:
      description: Get all complete tasks (is_complete = true) with only ID and title
      produces:
      - application/json
      responses:
        "200":
          description: 'data: []map[string]interface{}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Gagal mengambil data'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get incomplete tasks
      tags:
      - tasks
  /api/tasks/incomplete:
    get:
      description: Get all incomplete tasks (is_complete = false) with only ID and
        title
      produces:
      - application/json
      responses:
        "200":
          description: 'data: []map[string]interface{}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Gagal mengambil data'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get incomplete tasks
      tags:
      - tasks
  /api/tasks/search:
    get:
      description: Get all tasks for a specific search
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "400":
          description: 'error: Nama tidak val'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Gagal mengambil data'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get tasks by search
      tags:
      - calendar
  /api/tasks/weekly-stats:
    patch:
      consumes:
      - application/json
      description: Mark a task as complete or incomplete
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      - description: Completion status
        in: body
        name: completion
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Task completion status updated, todo_id: Todo ID'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Invalid task ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Tidak memiliki akses ke task ini'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Task tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Update failed'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get task completion
      tags:
      - tasks
...
  /api/todos/{id}/tasks:
    get:
      description: Get all tasks for a specific todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: []models.Task, id: Todo ID'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Todo ID tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Tidak memiliki akses ke todo ini'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Gagal mengambil data'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get tasks by todo ID
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task for a todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.Task'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Task berhasil dibuat, id: Todo ID, data: Task'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Tidak memiliki akses ke todo ini'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create task
      tags:
      - tasks
  /api/todos/{id}/tasks/{task_id}:
    delete:
      description: Delete a task
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Task berhasil dihapus'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Task ID tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Tidak memiliki akses ke task ini'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Task tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete task
      tags:
      - tasks
    get:
      description: Get a specific task by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: models.Task, id: Todo ID'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Task ID tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Tidak memiliki akses ke task ini'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Task tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get task by ID
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Update an existing task
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      - description: Task update data
        in: body
        name: task
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Task updated successfully, todo_id: Todo ID'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Invalid task ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Tidak memiliki akses ke task ini'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Task tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Update failed'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update task
      tags:
      - tasks
  /api/todos/{id}/tasks/{task_id}/complete:
    patch:
      consumes:
      - application/json
      description: Mark a task as complete or incomplete
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      - description: Completion status
        in: body
        name: completion
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Task completion status updated, todo_id: Todo ID'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Invalid task ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Tidak memiliki akses ke task ini'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Task tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Update failed'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update task completion
      tags:
      - tasks
...
swagger: "2.0"